Tried to override handleCollision
Collisions seem to only work under very specific circumstances right
now. If the ship isn't moving and the bullet contacts the ship on a
corner then it works fine. But it the ship is moving or the bullet hits
along a flat edge of the ship then the game will lock up like before.